Pause Menu now functions
The pause menu now actually pauses the game and can be unpaused. The Pause menu now appears with a fade in when the game is paused instead of always being active.  The buttons for returning to the level select and main menu now send debug messages but that is all they do for now.