Removed comments, added new weapon, moved ammo sprites to new folder
The basis for the sin shot works, but its tilt is off, but I do know why. Fixing it will be janky tbh.

Anyways, I was trying to figure out what's going on with the broken shoot and I realized its due to having two objects with the active player controller script.