level 4 crash fixed
Level 4 will now not crash when reloading the level or trying to get to the main menu or level select. The problem was the blood particles for the boss were being called after the eye was destroyed and Unity didn't like that. Now the Effect is called just like the explosion.